FWIP: History now includes optional descriptions for entries and expandable rows to show those descriptions
